requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
Pillow>=10.0.0
# face-recognition>=1.3.0  # Optional - requires CMake and dlib. MTCNN is used as primary detector.
mtcnn>=0.1.1
flask>=2.3.0
flask-cors>=4.0.0
jupyter>=1.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pandas>=2.0.0

